chore: add strict shared types and typed API
