+++ backend/types/http.ts
@@ -2,21 +2,4 @@
  * SPDX-License-Identifier: MIT
  */
 
-// backend/types/http.ts
-import type { RequestHandler, Request } from 'express';
-import type { ParamsDictionary } from 'express-serve-static-core';
-import type { ParsedQs } from 'qs';
-
-export type AuthedRequestHandler<
-  P extends ParamsDictionary = ParamsDictionary,
-  ResBody = any,
-  ReqBody = any,
-  ReqQuery = ParsedQs,
-> = RequestHandler<P, ResBody, ReqBody, ReqQuery>;
-
-export type AuthedRequest<
-  P extends ParamsDictionary = ParamsDictionary,
-  ResBody = any,
-  ReqBody = any,
-  ReqQuery = ParsedQs,
-> = Request<P, ResBody, ReqBody, ReqQuery>;
+export { AuthedRequestHandler, AuthedRequest, ApiResult } from '@shared/http';
