+++ frontend/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc -p tsconfig.json --noEmit && vite build",
+    "test": "vitest run",
     "preview": "vite preview",
     "lint": "eslint . --ext .ts,.tsx && tsc -p tsconfig.json --noEmit"
   },
