+++ shared/types/uploads.ts
@@ -1,0 +1,9 @@
+export interface UploadedFile {
+  id: string;
+  filename: string;
+  url: string;
+}
+
+export interface UploadResponse {
+  uploads: UploadedFile[];
+}
